requirements.txt
pytest
appium-python-client
selenium